toc rewrite/refactoring
