+++ lib/extensions/shmakowiki-toc/lib/shmakowikitoc.js
@@ -1,0 +1,6 @@
+var shmakowikitoc = require('./_shmakowikitoc.js');
+
+exports.transform = function(ast) {
+    console.log(ast[0]);
+    return shmakowikitoc.ShmakoWikiToc.match(ast, 'topLevel');
+};
