+++ lib/extensions/shmakowiki-toc/src/shmakowikitoc.js
@@ -1,0 +1,1 @@
+var api = require('../../../shmakowiki.js');
