+++ src/shmakowiki.ometajs
@@ -124,8 +124,7 @@ ShmakoWiki.extensions = {
         return OmetaHighlighter.matchAll(c, p);
     },
     'toc': function(c, p) {
-        var x = ShmakoWikiToc.parse(c, p, ShmakoWiki);
-        return [x.toc, x.ast];
+        return ShmakoWiki.matchAll(c, 'topLevel');
     }
 };
