+++ src/shmakowiki2html.ometajs
@@ -77,7 +77,11 @@ ShmakoWikiToHtml.extensions = {
             return '<pre><code class="'+ p + '">' +
                     ShmakoWikiToHtml._escape(c) +
                 '</code></pre>'
-        }
+        },
+    'toc': function(c, p) {
+        console.log(ShmakoWikiTocToHtml.tocToHtml(c[0], ShmakoWikiToHtml));
+        return ShmakoWikiTocToHtml.tocToHtml(c[0], ShmakoWikiToHtml);
+    }
 };
 ShmakoWikiToHtml.extensions.hl = ShmakoWikiToHtml.extensions.ohl;
